requirements.txt
boto3==1.42.63
botocore==1.42.63
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cssselect2==0.9.0
fonttools==4.62.1
html2image==2.0.7
idna==3.11
jmespath==1.1.0
lxml==6.0.2
numpy==2.4.2
pandas==3.0.1
pillow==12.1.1
pycparser==3.0
pydyf==0.12.1
PyMuPDF==1.27.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-pptx==1.0.2
reportlab==4.4.10
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
tinycss2==1.5.1
tinyhtml5==2.1.0
typing_extensions==4.15.0
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
websocket-client==1.9.0
xlsxwriter==3.2.9
zopfli==0.4.1
google-auth
google-auth-oauthlib
google-api-python-client
